+++ packages/vchart/src/chart/base-chart.ts
@@ -46,8 +46,7 @@ import {
   normalizeLayoutPaddingSpec,
   array,
   isTrueBrowser,
-  isString,
-  getActualColor
+  isString
 } from '../util';
 import { Stack } from './stack';
 import { BaseModel } from '../model/base-model';
@@ -59,7 +58,7 @@ import type { IParserOptions } from '@visactor/vdataset/es/parser';
 import type { IBoundsLike } from '@visactor/vutils';
 // eslint-disable-next-line no-duplicate-imports
 import { has, isFunction, isEmpty, getContainerSize } from '@visactor/vutils';
-import { getDataScheme } from '../theme/color-scheme/util';
+import { getActualColor, getDataScheme } from '../theme/color-scheme/util';
 import type { IGroupMark, IMorphConfig, IMark as IVGrammarMark, IView } from '@visactor/vgrammar';
 import { CompilableBase } from '../compile/compilable-base';
 import type { IStateInfo } from '../compile/mark/interface';
